Unify font type

+++ src/controls/checkbox.cc
@@ -10,7 +10,7 @@ void zgui::checkbox(const char* id, bool& value) noexcept
 	const int control_height = 8;
 	const int control_width = 8;
 
-	const int font = utils::misc::pop_font();
+	const unsigned long font = utils::misc::pop_font();
 
 	const vec2 cursor_pos = utils::misc::pop_cursor_pos();
 	const vec2 draw_pos{ context.window.position.x + cursor_pos.x, context.window.position.y + cursor_pos.y };
